Allow aliases prefix to be something else than `g`
There were a couple of reports before about aliases shadowing commands
in zimfw/zimfw#59 and zimfw/zimfw#139. And I was impressed to check how
many actually shadow existing commands using Ubuntu's command-not-found:

   'gb'  : apt install gb
   'gc'  : apt install graphviz
   'gcp' : apt install gcp
   'gcs' : apt install opengcs
   'gcv' : apt install scotch
   'gdc' : apt install gdc
   'gfm' : apt install gfm
   'gid' : apt install id-utils
   'gld' : apt install postfix-gld
   'gm'  : apt install graphicsmagick
   'gmt' : apt install gmt
   'gp'  : apt install pari-gp
   'gpa' : apt install gpa
   'gpp' : apt install gpp
   'gpt' : apt install gpt
   'grc' : apt install grc
   'gri' : apt install gri
   'gs'  : apt install ghostscript
   'gsl' : apt install generator-scripting-language
   'gss' : apt install libgss-dev
   'gt'  : apt install genometools
   'gtv' : apt install smpeg-gtv
   'gwc' : apt install geneweb
   'gwd' : apt install geneweb

The ability to customize the prefix fixes this issue. Picking `G` for
example is a good one (Vim's Fugitive uses upper case `G` in the
beginning of is commands, to mention another place where it's used).