Update log formats
Use fuller format with committer info instead of medium format, as that
is more complete, and info for author and committer seems to fit in one
line each.

Fix one-line medium format so it truncates the subject at 50, instead of
at the fixed column 60, as that was not good to graph logs.

Remove `--all` from graph log aliases, to increase their flexibility.
We're then able to use it just for the current or for a branch passed
as parameter (as we already could with the other log aliases).