Add Gbu alias, and rename git-branch-delete-gone
to a more generic git-branch-remote-tracking function.

I'd like to have an alias that also unsets the remote-tracking info for
remote branches that are gone, like a `GbG` + `Gbu`, but I very strongly
still want to keep the aliases to a maximum of 3 characters. Maybe if we
had some global aliases we could do something like `Gbg XGbx` for `GbG`,
and `Gbg XGbu`, where `Gbg` just lists the gone ones and the aliases
starting with `X` do `| xargs ...`