Add date to _git_log_oneline_medium_format
As I've been having to constantly switch to the fuller format just to
look up the date of a commit. No need for that.
Also remove `<>` from name, as that is used actually for the commiter
email instead of the name in the fuller format and will make the log
line 2 characters shorter  :- )